do not allow self delete

+++ frontend/src/utils/onDeleteUser.js
@@ -1,10 +1,19 @@
+import { verifyToken } from "#api/auth/verifyToken"
 import { deleteUserById } from "#api/users/deleteUserById"
 import { onRequestHandlerError } from "./onRequestHandlerError"
 import { onRequestHandlerNotification } from "./onRequestHandlerNotification"
 
 export const onDeleteUser = async (e, userId, setNotification, setCounter) => {
   e.preventDefault()
+
   try {
+    const authToken = await verifyToken()
+    if (authToken && authToken === userId) {
+      const message = "You can not delete yourself"
+      onRequestHandlerNotification(setNotification, message, setCounter)
+      return
+    }
+
     const response = await deleteUserById(userId)
 
     if (response && response.status === 200) {
